refactor : add authentication and navigation module
