+++ app/src/main/res/values/strings.xml
@@ -31,20 +31,7 @@
     <string name="dog_detail_weight_title">Peso</string>
     <string name="dog_detail_male_title">Macho</string>
     <string name="dog_detail_title">Detalles del perro</string>
-    <!--Authentication screen section-->
-    <string name="authentication_login_title">Iniciar sesión</string>
-    <string name="authentication_login_user_text_hint">Correo Electrónico</string>
-    <string name="authentication_login_password_text_hint">Contraseña</string>
-    <string name="authentication_login_button_text">Iniciar sesión</string>
-    <string name="authentication_login_buttonText_text">Registrarse</string>
-    <string name="authentication_login_buttonText_title">No tienes una cuenta?</string>
-    <string name="authentication_login_success_message">Sección iniciada correctamente</string>
-    <string name="authentication_log_out_title">Salir de la aplicación</string>
-    <string name="authentication_log_out_message">deseas cerrar la aplicación?</string>
-    <string name="authentication_password_regex_error">contraseña no valida, ingresa letras, numeros y
-    mayúsculas</string>
-    <string name="authentication_login_password_confirm_text_hint">Confirmar Contraseña</string>
-    <string name="authentication_create_user_screen_title">Crear Cuenta</string>
+
 
     <string name="add_dog_error">Error anadiendo perrito a la colección :(</string>
     <string name="permission_rationale_title">Aceptar los permisos</string>
