+++ authentication/src/main/java/com/kalex/dogescollection/authentication/RegexPatterns.kt
