+++ authentication/src/main/java/com/kalex/dogescollection/authentication/RegexValidationState.kt
