+++ authentication/src/main/java/com/kalex/dogescollection/authentication/model/AuthenticationUseCase.kt
