+++ authentication/src/main/res/drawable/ic_arrow_back_24.xml
