+++ settings.gradle
@@ -21,3 +21,5 @@ include ':app'
 include ':core'
 include ':camera'
 include ':doglist'
+include ':authentication'
+include ':navigation'
